+++ service/src/main/java/com/dashnet/dashNet/Task/TaskRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface TaskRepository extends CrudRepository<Task, Integer> {
+public interface TaskRepository extends CrudRepository<Task, Long> {
 	List<Task> findByTitleContaining(String title);
 	List<Task> findByDescriptionContaining(String description);
 	List<Task> findByTeamId(Long teamid);
